requirements.txt
networkx==3.5
openai==1.109.1
Pillow==11.3.0
python-dotenv==1.1.1
reportlab==4.4.4
